requirements.txt
asgiref==3.6.0
dj-database-url
Django==4.1.4
django-database-url==1.0.3
gunicorn==21.1.0
psycopg2==2.9.5
sqlparse==0.4.3
django-taggit==5.0.1
Pillow==10.2.0
pip==23.3.2
